refactor: add tag button
